Move SpringJobFactory to persistence
