+++ deposit/src/main/resources/service-context.xml
@@ -108,7 +108,7 @@
 		<property name="unavailableDelaySeconds" value="${unavailable.delay.seconds:60}"/>
 	</bean>
 
-	<bean id="jobFactory" class="edu.unc.lib.deposit.work.SpringJobFactory" />
+	<bean id="jobFactory" class="edu.unc.lib.dl.util.SpringJobFactory" />
 
 	<bean id="jobStatusFactory" class="edu.unc.lib.dl.util.JobStatusFactory" >
 		<property name="jedisPool" ref="jedisPool" />
