+++ persistence/src/main/java/edu/unc/lib/dl/util/SpringJobFactory.java
@@ -1,4 +1,4 @@
-package edu.unc.lib.deposit.work;
+package edu.unc.lib.dl.util;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
