+++ services/src/main/webapp/WEB-INF/service-context.xml
@@ -197,7 +197,7 @@
 		<constructor-arg ref="jesqueConfig" />
 	</bean>
 
-	<bean id="jobFactory" class="edu.unc.lib.deposit.work.SpringJobFactory" />
+	<bean id="jobFactory" class="edu.unc.lib.dl.util.SpringJobFactory" />
 	
 	<bean id="enhancementWorkerPool" class="net.greghaines.jesque.worker.WorkerPool">
 		<constructor-arg name="workerFactory">
